requirements.txt
scikit-learn>=1.4,<1.7
numpy>=1.26,<2.3
pandas>=2.1,<2.4
Pillow>=10.0,<11.2
pytesseract>=0.3.10,<0.4
joblib>=1.3,<1.5
streamlit>=1.30,<1.45
matplotlib>=3.8,<3.11
seaborn>=0.13,<0.14
